refactor(retrieval): carve me-query crate — hybrid search + query execution (Wave 2 #816 / S4, sub-PR 2)
Move the Query primitive (RRF merge, port-driven hybrid search, the
MemoryQuery builder, and MemoryEngine::{query, execute_query}'s bodies)
out of the facade into a new me-query crate as free functions over
MemoryCtx + ScopeTree + Reranker, per plan §552.

The facade's live retrieval path was already port-based
(port_hybrid_search over &dyn StorageBackend), so this is a clean
extraction with no runtime behavior change. Delete the superseded #631
Stage-C sync twin (hybrid_search + collect_candidates + their
rank_candidates ranking helper), verified dead-in-release and
consumed only by the sync↔port parity oracle test deleted alongside
it — the engine has driven the async path exclusively since the #631
cutover.

The archive-fallback augmentation in execute_query's search path
stays in the facade (MemoryEngine::execute_query, feature-gated):
search_archives_fallback reaches self.db_path/self.cold, state
MemoryCtx does not expose and that me-query has no dependency to
reach. me-query's execute_query ends where the original archive block
began; the facade delegate applies it as a post-step, gated on the
same has_search() && include_archives condition the original enforced
structurally.

Four bi-temporal/reranker helpers (validate_reranker_output,
passes_temporal_cutoff, fact_overlaps_period, fact_to_search_result)
lived in engine/mod.rs rather than engine/query.rs but are consumed
exclusively by the moved query bodies (verified: no other call site in
the workspace) — they move too, along with their proptest_temporal
property coverage.
