+++ memory-engine/lib/me-query/src/lib.rs
@@ -5,3 +5,15 @@
 //! `MemoryEngine::{query, execute_query}` are thin delegates over the free
 //! functions here.
 #![cfg_attr(test, allow(clippy::unwrap_used))]
+
+pub mod execute;
+pub mod hybrid;
+pub mod query;
+
+// `execute::query` (the free fn) is deliberately NOT flat-re-exported here: it would
+// collide with the `query` module (holding `MemoryQuery`) at the crate root, exactly
+// the naming tension the facade's original `search::query`/`engine::query` module
+// split existed to avoid. Callers reach it as `me_query::execute::query(...)`.
+pub use execute::execute_query;
+pub use hybrid::{port_hybrid_search, rrf_merge};
+pub use query::MemoryQuery;
