+++ memory-engine/lib/memory-engine/Cargo.toml
@@ -47,6 +47,10 @@ me-resolve = { path = "../me-resolve" }
 # sub-PR 1). `MemoryEngine::{ingest, add_fact, add_fact_precomputed,
 # add_facts_batch, add_facts_batch_partial}` delegate to it directly.
 me-ingest = { path = "../me-ingest" }
+# L3 Query primitive: hybrid FTS5 + vector + RRF retrieval (Wave 2 #816 / S4,
+# sub-PR 2). `MemoryEngine::{query, execute_query}` delegate to it directly; the
+# facade's `search` module also re-exports its `hybrid`/`query` submodules.
+me-query = { path = "../me-query" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 chrono = { version = "0.4", features = ["serde"] }
@@ -210,6 +214,7 @@ archive = [
     "me-storage/archive",
     "me-index/archive",
     "me-backend-sqlite/archive",
+    "me-query/archive",
 ]
 # The default in-process SQLite backend. NOTE (#633): this is a *marker* feature — it
 # does NOT yet gate `mod sqlite` or `rusqlite`. The engine open path hardcodes
